producer-traffic/requirements.txt
pymongo
python-dotenv
requests
boto3